requirements.txt
discord.py>=2.5.1
python-dotenv>=1.0.0
pytest>=8.0.0
flask==3.1.0
flask-cors==5.0.0
psutil==6.1.0
gunicorn==23.0.0
